Moved AST-components into their own subpackage.
