+++ server/src/main/java/de/uftos/dto/ucdl/ast/AbstractSyntaxTreeDto.java
@@ -1,4 +1,6 @@
-package de.uftos.dto.ucdl;
+package de.uftos.dto.ucdl.ast;
+
+import de.uftos.dto.ucdl.UcdlToken;
 
 /**
  * The abstract syntax tree describing the structure of a constraint definition.
